requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib[argon2]
python-multipart